+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # History
 
+- 2026-07-04:
+  - 🆕 added `getItemMetadata` to API item service and buttons to generate item's title and description when environment `hasMetadataBuilders` is set to true (`env.js`): `@myrmidon/cadmus-api`, `@myrmidon/cadmus-item-editor`.
 - 2026-07-02:
   - display facet label instead of its ID in items list.
   - updated Angular and packages.